fix(CoreUIHelpers): split functions
